// src/domain/calidad/models/CalidadPorcentajeGeneralDanosInsert.ts
export interface CalidadPorcentajeGeneralDanosInsert {
  IdLote: number;
  Trips: string;
  Rona: string;
  Golpe: string;
  Rozadura: string;
  Varicela: string;
  Clavo: string;
  Alga: string;
  Escama: string;
  Granizo: string;
  Gusano: string;
  Quemadura: string;
  Sunblotch: string;
  IdHuertasPesoCalibre: number;
}